requirements.txt
APScheduler==3.8.0
asgiref==3.4.1
autopep8==1.5.7
click==8.0.1
Flask==2.0.1
Flask-APScheduler==1.12.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
flask-migrate==3.1.0
greenlet==1.1.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
pycodestyle==2.7.0
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.1
scipy==1.7.1
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
typing-extensions==3.10.0.2
tzlocal==2.1
Werkzeug==2.0.1
WTForms==2.3.3
zipp==3.6.0
psycopg2==2.8.6
gunicorn